checkers: sh a.c gate now enforces every checker, all pass, tok headroom line, -469 a-tok
Sean 2026-09-05: "fix the checker errors neg tok" · "when i sh a.c i want to be told how close we are to tok killing limit add that in".

- sh a.c prints `tok N/CAP (D left|over)` on every build, before the TOK KILL verdict.
- gate bug: `ls "$T"/[0-9].f "$T"/1[0-9].f` exits 2 whenever one glob is unmatched, so a failing checker 1-9 was only acted on if a checker >=10 also failed: the checkers never gated a build. Now `ls "$T"/*.f`.
- checker 4 (gcc): -Werror=<flag> per flag; blanket -Werror turned the #if 0 python half's apostrophes ("missing terminating ' character") into errors on every run.
- checker 5 (gcc -fanalyzer -fsyntax-only) deleted: -fsyntax-only never runs the analyzer (0 findings on a planted leak, 2 with -c).
- clang -Weverything: ~200 errors cleared: reserved _identifiers renamed (T0, serve.c helpers, ldcmp, bkfile, spk, doctree), shadows renamed, comma-expression returns braced, sign/const/cast fixes, unused sync_pane deleted, tm_ensure_conf bind loop tabled; 3 -Wno flags (format-nonliteral, implicit-int-float-conversion, overlength-strings).
- cppcheck: non-ASCII in a.c's shell comments ASCII-fied.
- book.c bk_mid: `(*e&0xC0)==(char)0x80` never matched (int vs signed char), so the middle-elision could split a UTF-8 sequence; now `==0x80`.

Verified: clang-23 -Weverything -Werror -fsyntax-only 0 errors; a.c's own _checkers (honest glob) ALL PASSED; 42 HTTP routes byte-identical base/new (a_base=HEAD, interleaved base/new/base); `a i` piped identical (1307 rows); tmux.conf generation byte-identical (5258 B); smoke ok.
Still over cap: a.c+lib 231,438 vs 230,000 (1,438 over); pure I/O-preserving crunch is exhausted, the rest is a deletion for Sean to name.

+++ lib/work.c
@@ -5,7 +5,7 @@ static int cmd_w(int c,char**v){AB;init_paths();
     char*h=readf(f,NULL);if(!h)return 1;
     for(char*l=h;*l&&n<32;l++){char*p=strchr(l,'\n');if(p)*p=0;if(*l){char*at=strchr(l,'@');if(at){*at=0;ht[n]=at+1;}hl[n++]=l;}l=p?p:l+strlen(l);}
     char ds[12];strftime(ds,12,"%F",localtime(&(time_t){time(0)}));
-    snprintf(f+dl,P-dl,"/%s_%s.txt",ds,DEV);
+    snprintf(f+dl,(size_t)(P-dl),"/%s_%s.txt",ds,DEV);
     char*dn=readf(f,NULL);
     if(c>2&&!strcmp(v[2],"check")){perf_disarm();char wf[P];snprintf(wf,P,"%s/local/wcheck_%s.txt",AROOT,ds);char*wd=readf(wf,NULL);
         struct tm*T=localtime(&(time_t){time(0)});int now=T->tm_hour*60+T->tm_min,hh,mm;char cm[B];
